Bug 1289258 - Part 3: Convert action test to Mocha. r=bgrins
