--- devtools/client/webconsole/new-console-output/test/actions/xpcshell.ini
@@ -1,7 +1,0 @@
-[DEFAULT]
-tags = devtools devtools-webconsole
-head = head.js
-tail =
-firefox-appdir = browser
-
-[test_messages.js]
